clothing_store/requirements.txt
Django==5.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
django-simpleui==2023.12.30
Pillow==10.1.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn==21.2.0
celery==5.3.4
redis==5.0.1
whitenoise==6.6.0
dj-database-url==2.1.0
